reverting alchemical_network file renaming. this will be part of a different PR
