+++ gufe/network.py
